+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### 2.0.1 - July 24, 2022
   - [BUGFIX] Context menu wasn't opening in HUD after exhausting a source - had to reopen HUD.
+  - [BUGFIX] Making Dancing Lights respond to configured settings.
 
 ### 2.0.0 - July 23, 2022
    - [BREAKING] (Lupestro) This release supports v10 of Foundry, and not v0.7, v0.8, or v9.